Task_Management_Api/requirements.txt
Django
django-filter
djangorestframework
djangorestframework-simplejwt
drf-spectacular
drf-spectacular-sidecar
gunicorn
psycopg2-binary
python-dotenv
python-decouple
uvicorn
whitenoise
requests
faker